Fix embassy-usb panic from configuration
embassy-usb expects a particular config if composite_with_iads is set,
(the default) and the previous settings would cause an assertion to fail.

According to the Nordic documentation:

> IAD should not be required for non-composite CDC ACM device,
> but Windows 7 fails to properly enumerate without it.
> Enable if you want CDC ACM to work with Windows 7.

https://docs.nordicsemi.com/bundle/ncs-1.9.1/page/kconfig/CONFIG_CDC_ACM_IAD.html

Ideally we'd just use the standard USB serial class code (0x02) rather
than the one for IADs (0xEF), and Windows 7 is now out of extended
support, but this is still the default in embassy-usb and it makes sense
just to track what it does.