Controller: Move buffers into the controller struct
The array-of-buffers design has been kept because it's important to keep
the buffers at the same memory location, rather than (as I experimented
with) mem::swap-ing .active_buffer and .other_buffer fields.

This is because the logger task, which writes the buffer contents out
over USB, takes a reference to the buffer it's flushing. If you
mem::swap the fields, it ends up replacing the the &mut LogBuffer you're
holding.

*This* can happen because the controller is stored in a static mut, so
we cannot control who can take a &mut to it. And even if we replace the
static mut with interior mutability, we have to reckon with UnsafeCell
and mutation while the USB logger task is suspended .await-ing a packet
being written out (and defmt calls might occur in other tasks).

I've replaced the BufferIndex enum with just a usize that is xor-ed with
1 to produce either the 0 or 1 index. The BufferIndex's swap method
(which returns the alternate index without mutating self) was the cause
of the frame corruption problem (not storing the swapped index), really
because of accidental misuse, but it seemed likely that the extra layer
wasn't helpful in getting the swapping correct as opposed to a plain
indexing usize.