Controller: reset both buffers on disable
This in effect "clears" any leftover defmt frames from before the
controller was disabled. This is important because the controller is
disabled when it is discovered that the USB endpoint is disabled.

This may result in the consumer not receiving the end of a frame, but
this is not considered a problem because the consumer is the one that
has disabled the USB endpoint (eg, quit defmt-print).

Resetting both buffers ensures that there is no data to be sent when the
USB endpoint is reconnected, so no stale, potentially partial and
appearing "corrupt" data is transmitted.